+++ tchart/renderers.py
@@ -95,3 +95,42 @@ class SharpRenderer(BoxRenderer):
         u'x',  # 2
         u'#',  # 3
     )
+
+
+class DrawilleRenderer(ChartRenderer):
+    @property
+    def char_resolution_horizontal(self):
+        return 2
+
+    @property
+    def char_resolution_vertical(self):
+        return 4
+
+    def render(self, width, height, values):
+        """
+        :type width: int
+        :type height: int
+        :type values: list[int or float]
+        :rtype: list[str or unicode]
+        """
+        from drawille import Canvas
+
+        vertical_chart_resolution = height * self.char_resolution_vertical
+        horizontal_chart_resolution = width * self.char_resolution_horizontal
+
+        canvas = Canvas()
+        x = 0
+        for value in values:
+            for y in range(1, int(round(value, 0)) + 1):
+                canvas.set(x, vertical_chart_resolution - y)
+            x += 1
+
+        rows = canvas.rows(min_x=0, min_y=0, max_x=horizontal_chart_resolution, max_y=vertical_chart_resolution)
+
+        if not rows:
+            rows = [u'' for _ in range(height)]
+
+        for i, row in enumerate(rows):
+            rows[i] = u'{0}'.format(row.ljust(width))
+
+        return rows